+++ rascal-lsp/src/main/java/org/rascalmpl/vscode/lsp/parametric/ILanguageContributions.java
@@ -72,22 +72,22 @@ public interface ILanguageContributions {
     public CompletableFuture<IList> parseCodeActions(String command);
     public CompletableFuture<IConstructor> parseCallHierarchyData(String data);
 
-    public CompletableFuture<Boolean> hasAnalysis();
-    public CompletableFuture<Boolean> hasBuild();
-    public CompletableFuture<Boolean> hasDocumentSymbol();
-    public CompletableFuture<Boolean> hasCodeLens();
-    public CompletableFuture<Boolean> hasInlayHint();
-    public CompletableFuture<Boolean> hasRename();
-    public CompletableFuture<Boolean> hasExecution();
-    public CompletableFuture<Boolean> hasHover();
-    public CompletableFuture<Boolean> hasDefinition();
-    public CompletableFuture<Boolean> hasReferences();
-    public CompletableFuture<Boolean> hasImplementation();
-    public CompletableFuture<Boolean> hasCodeAction();
-    public CompletableFuture<Boolean> hasDidRenameFiles();
-    public CompletableFuture<Boolean> hasSelectionRange();
-    public CompletableFuture<Boolean> hasCallHierarchy();
-    public CompletableFuture<Boolean> hasCompletion();
+    public CompletableFuture<Boolean> providesAnalysis();
+    public CompletableFuture<Boolean> providesBuild();
+    public CompletableFuture<Boolean> providesDocumentSymbol();
+    public CompletableFuture<Boolean> providesCodeLens();
+    public CompletableFuture<Boolean> providesInlayHint();
+    public CompletableFuture<Boolean> providesRename();
+    public CompletableFuture<Boolean> providesExecution();
+    public CompletableFuture<Boolean> providesHover();
+    public CompletableFuture<Boolean> providesDefinition();
+    public CompletableFuture<Boolean> providesReferences();
+    public CompletableFuture<Boolean> providesImplementation();
+    public CompletableFuture<Boolean> providesCodeAction();
+    public CompletableFuture<Boolean> providesDidRenameFiles();
+    public CompletableFuture<Boolean> providesSelectionRange();
+    public CompletableFuture<Boolean> providesCallHierarchy();
+    public CompletableFuture<Boolean> providesCompletion();
 
     public CompletableFuture<Boolean> specialCaseHighlighting();